+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Change Cheese to be installed as a Flatpak.
 - Change Discord to be installed as a Flatpak.
 - Change Gwenview to be installed as a Flatpak.
+- Change KeePassXC to be installed as a Flatpak.
 
 ## 3.0.1